Move eval function and finishing touches
